# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NEXUS — API del Sistema Principal
  description: |
    API del módulo `SP`. **Toda operación exige un token de acceso**
    salvo el inicio, la renovación y el cierre de sesión.

    Para probar cualquier endpoint: obtenga el token con
    `POST /api/v1/auth/login`, pulse **Authorize** y pegue el valor de
    `accessToken` — sin el prefijo `Bearer`, que Swagger añade solo.

    El token vive **quince minutos** y no es revocable: cuando expire,
    renuévelo con `POST /api/v1/auth/refresh` y vuelva a autorizar.
  version: v1
servers:
- url: /
  description: Este mismo servidor
security:
- bearerAuth: []
...
paths:
...
  /api/v1/users:
    get:
      tags:
      - Usuarios
      summary: Consultar personas
      description: |
        Listado paginado con filtros, búsqueda y ordenamiento.

        **El orden por defecto es `lastName,asc`** y no el nombre de usuario:
        esta es la lista desde la que se administra el acceso, y quien la mira
        busca a alguien por su apellido.

        Solo se puede ordenar por la lista blanca —`username`, `email`,
        `firstName`, `lastName`, `status`, `createdAt`, `updatedAt`—. **No se
        admite ordenar por ningún campo de la credencial**: ordenar por la marca
        de cambio obligatorio produciría la lista de quien no ha cambiado su
        contraseña inicial.

        **`roleId`, `membershipId` y `countryId` no se validan contra su
        catálogo.** Un filtro por algo inexistente devuelve la colección vacía y
        no es un error.

        **`countryId` NO se acota a países activos**, y va contra la intuición:
        filtrar solo por los activos convertiría desactivar un país en una forma
        de **esconder a su gente**, y este listado es justamente la herramienta
        con la que se busca a quien quedó dentro para moverlo. Quien esté en un
        país retirado aparece con normalidad.

        Cada fila trae su `country` **resuelto y nunca nulo** — es el único
        objeto anidado de la fila del que se puede decir eso: `roles` puede venir
        vacía y `membership` y `deletedAt` pueden venir nulos.

        **El listado NO publica el documento ni los datos de contacto**, y no se
        puede buscar por ellos. Es deliberado: exponer un documento en un listado
        paginado alcanza a mucha más gente que devolverlo en un detalle. Quien
        necesite el documento de una persona usa `GET /api/v1/users/{id}`.

        La búsqueda va sobre nombre de usuario, correo y nombre completo,
        **sin distinguir acentos ni mayúsculas**, y por fragmento.

        `membership` **no es nula cuando está vencida**: vencer no es lo mismo
        que no tener. `current` dice cuál de los dos casos es.

        Un filtro sin coincidencias devuelve `200` con la colección vacía. No
        hay `404` ni `422`.
      operationId: listar
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: sort
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: roleId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: membershipId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: countryId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: includeDeleted
        in: query
        required: false
        schema:
          type: boolean
      responses:
        "200":
          description: Página de personas.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponse"
        "400":
          description: "Paginación fuera de límites (`VAL-001`, `VAL-002`), campo\
            \ de ordenamiento no admitido (`VAL-003`), o estado fuera de su dominio\
            \ (`VAL-004`). Se devuelven **juntos**"
        "401":
          description: Token ausente o inválido (`AUTH-001`)
        "403":
          description: Autenticado sin `users:read` (`AUTH-002`)
        "500":
          description: Fallo no controlado (`ERR-500`)
    post:
      tags:
      - Usuarios
      summary: Registrar una persona
      description: |
        Registra una persona con sus roles iniciales.

        La cuenta nace **activa** y **marcada para cambio obligatorio de
        contraseña**: quien prepara el alta conoce la credencial, y esa ventana
        se cierra en el primer inicio de sesión. Ni el estado ni la marca se
        envían; enviarlos devuelve `400`.
        **`membershipId` es OPCIONAL, y el superior es condicional.** Desde el
        05-09-2026 toda persona nace con nivel: si no se indica ninguno, se le
        concede el de arranque (`BECA`). Ya no hay rechazo por indicarlo sin rol
        de consumidor ni por omitirlo teniéndolo — esa exigencia murió con
        `RN-SP-013`.

        El rol de vendedor y el superior comercial **sí siguen siendo
        condicionales en los dos sentidos**: indicar uno sin el otro devuelve
        `409`, no se ignora.

        El superior debe portar el **rol padre inmediato** del rol vendedor de
        mayor rango de la persona, y estar activo.

        ## Identidad, contacto y país

        **`countryId`, `documentTypeId`, `documentNumber` y `phone` son
        obligatorios**, y a diferencia de la membresía y el superior **no
        dependen de qué roles se concedan**: su ausencia es `400`, nunca un
        `409` condicional.

        **`companyPhone`, `addressLine1`, `addressLine2` y `city` son
        opcionales.** Exigir una dirección postal a un funcionario interno
        bloquearía su alta sin que nadie la necesite. `city` es **texto
        libre**: no hay catálogo de ciudades.

        **El tipo de documento sale de `GET /api/v1/document-types`, y ese
        catálogo contiene SOLO documentos de persona mayor de edad.** Ahí está
        la validación de mayoría de edad del sistema entero: no hay ninguna
        marca que diga cuáles acreditan y cuáles no, porque **los que no
        acreditan no están**. Registrar a un menor no se rechaza — no se puede
        expresar, porque no hay identificador que enviar. Un desplegable
        alimentado por ese catálogo no puede ofrecer una opción que el alta vaya
        a rechazar.

        **El par tipo + número es único entre TODAS las personas, incluidas las
        eliminadas** (`RN-SP-035`), igual que el nombre de usuario y el correo.
        El `409` que produce **no dice de quién es** el documento, ni si esa
        persona sigue vigente.

        El número de documento se guarda **recortado y en mayúsculas**, y
        los **dos teléfonos** —`phone`, el personal, y `companyPhone`, el de la
        empresa— **normalizados a dígitos con un `+` opcional**, fuera espacios,
        guiones y paréntesis. La dirección y la ciudad solo se recortan.

        **`companyPhone` es OPCIONAL y `phone` no** (`RN-SP-037`): exigir un
        teléfono de empresa bloquearía el alta de todo el que no tenga una. Sale
        **presente y en nulo** cuando no se declara, nunca ausente.

        **Ningún teléfono se valida contra el país**: eso exigiría un catálogo de
        prefijos que no existe.

        La respuesta agrupa `country`, `document` y `contact` en tres objetos, y
        devuelve el **tipo de documento resuelto** —abreviación y nombre— para
        que no haga falta una segunda llamada al catálogo, que además exige otro
        permiso.

        La contraseña no se recorta: un espacio al principio o al final es parte
        de ella.
      operationId: registrar
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
        required: true
      responses:
        "201":
          description: Persona registrada. `Location` apunta a su detalle.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: "Formato u obligatoriedad incumplidos —incluidos país, tipo\
            \ y número de documento y teléfono, que son obligatorios (`VAL-014` a\
            \ `VAL-017`)—, contraseña que no cumple la política, o campo no admitido\
            \ en el cuerpo"
        "401":
          description: Token ausente o inválido (`AUTH-001`)
        "403":
          description: Autenticado sin el permiso de creación de usuarios (`AUTH-002`)
        "409":
          description: "Identidad ya en uso (`RN-SP-016`), documento ya registrado\
            \ por otra persona —vigente o eliminada— o tipo de documento inactivo\
            \ (`RN-SP-035`), país inactivo (`RN-SP-034`), rol que excede los privilegios\
            \ del actor (`RN-SEG-010`), vendedor sin superior o al revés (`RN-SP-019`),\
            \ o superior que no porta el rol padre (`RN-SP-020`)"
        "422":
          description: "Algún rol no existe o no está activo (`EX-003`), el país no\
            \ existe (`EX-009`) o el tipo de documento no existe (`EX-010`) — referencias\
            \ que no resuelven, frente al `409` de las que resuelven y una regla rechaza"
        "500":
          description: Fallo no controlado (`ERR-500`)
...
  /api/v1/users/{id}/deletion:
    post:
      tags:
      - Usuarios
      summary: Eliminar a una persona
      description: |
        Eliminación **lógica** y con motivo declarado. `POST` sobre un
        subrecurso y no `DELETE` con cuerpo: RFC 9110 no define semántica para
        el cuerpo de un `DELETE` y un intermediario puede descartarlo,
        convirtiendo la petición en un rechazo por motivo ausente que el actor
        no puede entender. Y tampoco por *query string*, o el motivo acabaría en
        los registros de acceso de los proxies.

        **El motivo se verifica el primero de todo**, antes incluso de saber si
        la persona existe.

        La operación **retira sus roles y su membresía**, **cierra** su
        asignación de superior —cerrarla, nunca borrarla: es historial de
        mando— y **revoca sus sesiones**, todo en la misma transacción y con la
        misma marca de tiempo.

        **El estado NO se toca**: se conserva como estaba para que el registro
        de eliminación diga en qué situación estaba la persona cuando se la
        eliminó.

        El `404` **no distingue** «nunca existió» de «ya estaba eliminada».
      operationId: eliminar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteUserRequest"
        required: true
      responses:
        "204":
          description: Persona eliminada.
        "400":
          description: "Motivo ausente o vacío (`VAL-001`), o campo desconocido"
        "401":
          description: Token ausente o inválido (`AUTH-001`)
        "403":
          description: "Autenticado sin `users:delete` (`AUTH-002`), o es la cuenta\
            \ del propio actor (`RN-SP-017`)"
        "404":
          description: "No existe, o ya estaba eliminada (`EX-004`)"
        "409":
          description: "Es el último superadministrador activo (`RN-SP-001`), o tiene\
            \ personas a cargo (`RN-SP-022`)"
        "500":
          description: Fallo no controlado (`ERR-500`)
...
  /api/v1/user-commission-rates/{id}/products/{productId}/deletion:
    post:
      tags:
      - Comisiones
      summary: Retirar la asociación de una tasa personalizada con un producto
      description: |
        **Deja de regir ahí, y la tasa sigue viva.** Puede volver a asociarse a
        ese producto o a otro; retirarla es otra operación.

        **El motivo es obligatorio.** La fila se borra de verdad —una asociación
        es configuración vigente, no un hecho del pasado— de modo que este texto
        es **el único sitio** donde quedará escrito por qué esa persona dejó de
        tener su excepción en ese producto.

        **`404` y no `409` si ya no estaba**: con el borrado físico no queda nada
        que distinga «nunca existió» de «ya se borró», y un `409` afirmaría algo
        que no se sabe.
      operationId: desasociar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DissociateProductRequest"
      responses:
        "204":
          description: Desasociada
        "400":
          description: "Motivo ausente, en blanco o demasiado largo"
        "403":
          description: Sin permiso
        "404":
          description: Esa tasa no está asociada a ese producto
  /api/v1/user-commission-rates/{id}/deletion:
    post:
      tags:
      - Comisiones
      summary: Retirar una tasa personalizada
      description: |
        Retira una tasa que **no debió existir**, con **motivo obligatorio**.

        **Retirar no es cerrar la vigencia**: se cierra lo que dejó de regir, se
        retira lo que fue un error. Por eso el retiro **no toca la vigencia** — el
        registro de eliminación debe poder decir qué periodo cubría lo retirado.

        Los días que ocupaba **quedan libres**, de modo que puede declararse otra
        tasa que los cubra.

        **No es idempotente**: retirar dos veces devuelve `409`.
      operationId: retirar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteCommissionRateRequest"
      responses:
        "204":
          description: Tasa retirada
        "400":
          description: "Motivo ausente, en blanco o demasiado largo"
        "403":
          description: Sin permiso
        "404":
          description: La tasa no existe
        "409":
          description: Ya estaba retirada
...
  /api/v1/commission-rates/{id}/products/{productId}/deletion:
    post:
      tags:
      - Comisiones
      summary: Retirar la asociación de la tasa con un producto
      description: |
        **Es la única forma de dejar de pagar sin retirar la tasa**: sigue en el
        catálogo y disponible para otros productos.

        **El borrado es físico** y la tabla no tiene retiro lógico, de modo que
        **el registro de eliminación es lo único que queda** de que esa tasa rigió
        alguna vez sobre ese producto. De ahí el **motivo obligatorio**.

        Si esa tasa no está asociada a ese producto devuelve `404` — y no `409`,
        porque al no quedar rastro no puede distinguirse «nunca existió» de «ya se
        borró».
      operationId: desasociar_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DissociateProductRequest"
      responses:
        "204":
          description: Asociación retirada
        "400":
          description: "Motivo ausente, en blanco o demasiado largo"
        "403":
          description: Sin permiso
        "404":
          description: Esa tasa no está asociada a ese producto
...
components:
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        status:
          type: string
        mustChangePassword:
          type: boolean
        roles:
          type: array
          items:
            $ref: "#/components/schemas/RoleRef"
        country:
          $ref: "#/components/schemas/CountryRef"
        document:
          $ref: "#/components/schemas/DocumentRef"
        contact:
          $ref: "#/components/schemas/ContactRef"
        membership:
          $ref: "#/components/schemas/MembershipRef"
        supervisor:
          $ref: "#/components/schemas/SupervisorRef"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    DeleteUserRequest:
      type: object
      properties:
        reason:
          type: string
...
    DissociateProductRequest:
      type: object
      properties:
        reason:
          type: string
    DeleteCommissionRateRequest:
      type: object
      properties:
        reason:
          type: string
...
    PageResponse:
      type: object
      properties:
        content:
          type: array
          items: {}
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalIsExact:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      description: "Token de acceso emitido por `POST /api/v1/auth/login`. Pegue solo\
        \ el valor de `accessToken`: el prefijo `Bearer` lo añade la interfaz."
      in: header
      scheme: bearer
      bearerFormat: JWT